Python/Django/s3cms/requirements.txt
boto==2.38.0
django-storages==1.1.8
Pillow==2.9.0
